Create DocumentModel trait to enable MongoDB on any 3rd party model class
