+++ src/Relations/MorphToMany.php
@@ -77,7 +77,7 @@ class MorphToMany extends EloquentMorphToMany
     protected function setWhere()
     {
         if ($this->getInverse()) {
-            if ($this->parent instanceof \MongoDB\Laravel\Eloquent\Model) {
+            if (\MongoDB\Laravel\Eloquent\Model::isDocumentModel($this->parent)) {
                 $ids = $this->extractIds((array) $this->parent->{$this->table});
 
                 $this->query->whereIn($this->relatedKey, $ids);
@@ -86,7 +86,7 @@ class MorphToMany extends EloquentMorphToMany
                     ->whereIn($this->foreignPivotKey, (array) $this->parent->{$this->parentKey});
             }
         } else {
-            match ($this->parent instanceof \MongoDB\Laravel\Eloquent\Model) {
+            match (\MongoDB\Laravel\Eloquent\Model::isDocumentModel($this->parent)) {
                 true => $this->query->whereIn($this->relatedKey, (array) $this->parent->{$this->relatedPivotKey}),
                 false => $this->query
                     ->whereIn($this->getQualifiedForeignPivotKeyName(), (array) $this->parent->{$this->parentKey}),
@@ -140,7 +140,7 @@ class MorphToMany extends EloquentMorphToMany
         // in this joining table. We'll spin through the given IDs, checking to see
         // if they exist in the array of current ones, and if not we will insert.
         if ($this->getInverse()) {
-            $current = match ($this->parent instanceof \MongoDB\Laravel\Eloquent\Model) {
+            $current = match (\MongoDB\Laravel\Eloquent\Model::isDocumentModel($this->parent)) {
                 true => $this->parent->{$this->table} ?: [],
                 false => $this->parent->{$this->relationName} ?: [],
             };
@@ -151,7 +151,7 @@ class MorphToMany extends EloquentMorphToMany
                 $current = $this->extractIds($current);
             }
         } else {
-            $current = match ($this->parent instanceof \MongoDB\Laravel\Eloquent\Model) {
+            $current = match (\MongoDB\Laravel\Eloquent\Model::isDocumentModel($this->parent)) {
                 true => $this->parent->{$this->relatedPivotKey} ?: [],
                 false => $this->parent->{$this->relationName} ?: [],
             };
@@ -213,7 +213,7 @@ class MorphToMany extends EloquentMorphToMany
 
             if ($this->getInverse()) {
                 // Attach the new ids to the parent model.
-                if ($this->parent instanceof \MongoDB\Laravel\Eloquent\Model) {
+                if (\MongoDB\Laravel\Eloquent\Model::isDocumentModel($this->parent)) {
                     $this->parent->push($this->table, [
                         [
                             $this->relatedPivotKey => $model->{$this->relatedKey},
@@ -236,7 +236,7 @@ class MorphToMany extends EloquentMorphToMany
                 ], true);
 
                 // Attach the new ids to the parent model.
-                if ($this->parent instanceof \MongoDB\Laravel\Eloquent\Model) {
+                if (\MongoDB\Laravel\Eloquent\Model::isDocumentModel($this->parent)) {
                     $this->parent->push($this->relatedPivotKey, (array) $id, true);
                 } else {
                     $this->addIdToParentRelationData($id);
@@ -257,7 +257,7 @@ class MorphToMany extends EloquentMorphToMany
                 $query->push($this->foreignPivotKey, $this->parent->{$this->parentKey});
 
                 // Attach the new ids to the parent model.
-                if ($this->parent instanceof \MongoDB\Laravel\Eloquent\Model) {
+                if (\MongoDB\Laravel\Eloquent\Model::isDocumentModel($this->parent)) {
                     foreach ($id as $item) {
                         $this->parent->push($this->table, [
                             [
@@ -281,7 +281,7 @@ class MorphToMany extends EloquentMorphToMany
                 ], true);
 
                 // Attach the new ids to the parent model.
-                if ($this->parent instanceof \MongoDB\Laravel\Eloquent\Model) {
+                if (\MongoDB\Laravel\Eloquent\Model::isDocumentModel($this->parent)) {
                     $this->parent->push($this->relatedPivotKey, $id, true);
                 } else {
                     foreach ($id as $item) {
@@ -324,7 +324,7 @@ class MorphToMany extends EloquentMorphToMany
                 ];
             }
 
-            if ($this->parent instanceof \MongoDB\Laravel\Eloquent\Model) {
+            if (\MongoDB\Laravel\Eloquent\Model::isDocumentModel($this->parent)) {
                 $this->parent->pull($this->table, $data);
             } else {
                 $value = $this->parent->{$this->relationName}
@@ -341,7 +341,7 @@ class MorphToMany extends EloquentMorphToMany
             $query->pull($this->foreignPivotKey, $this->parent->{$this->parentKey});
         } else {
             // Remove the relation from the parent.
-            if ($this->parent instanceof \MongoDB\Laravel\Eloquent\Model) {
+            if (\MongoDB\Laravel\Eloquent\Model::isDocumentModel($this->parent)) {
                 $this->parent->pull($this->relatedPivotKey, $ids);
             } else {
                 $value = $this->parent->{$this->relationName}
